Move TileTransform into src
